+++ app/admin/components/cmsDirective.js
@@ -2,36 +2,35 @@
     'use strict';
 
     angular.module('app.admin')
-        .controller('CmsController', ['$log', 'commonService', 'authService', 'API', function ($log, commonService, authService, API) {
+        .controller('CmsController', ['$log', 'commonService', function ($log, commonService) {
             var vm = this;
 
-            vm.getDownload = getDownload();
+            vm.getDownload = getDownload;
 
             activate();
 
             ////////////////////////////////////////////////////////////////////
 
             function activate () {
-                vm.API = API;
-                vm.API_KEY = authService.getApiKey();
-                //vm.loadAnnouncements();
-                //vm.isChplAdmin = authService.isChplAdmin();
-                //vm.errorMessage = '';
+                vm.filename = 'CMS_Ids_' + new Date().getTime() + '.csv';
+                vm.csvHeader = ['CMS ID', 'Creation Date'];
+                vm.csvColumnOrder = ['certificationId', 'created'];
+                vm.isReady = false;
+                vm.getDownload();
             }
 
             function getDownload () {
+                commonService.getCmsDownload()
+                    .then(function (result) {
+                        for (var i = 0; i < result.length; i++) {
+                            result[i].created = new Date(result[i].created).toISOString().substring(0, 10);
+                        }
+                        vm.cmsArray = result
+                        vm.isReady = true;
+                    }, function (error) {
+                        $log.debug('error in app.admin.cmsController.getDownload', error);
+                    });
             }
-
-            /*
-            function loadAnnouncements () {
-            	commonService.getAnnouncements(true)
-            		.then (function (result) {
-            			vm.announcements = result.announcements;
-            		}, function (error) {
-            			$log.debug('error in app.admin.announcement.controller.loadAnnouncements', error);
-            		});
-            }
-*/
         }])
         .directive('aiCmsManagement', [function () {
             return {