+++ app/app.js
@@ -2,7 +2,7 @@
     'use strict';
 
     angular.module('app', ['ngRoute', 'angular-loading-bar', 'ngAnimate', 'angulartics', 'angulartics.google.analytics', 'angulartics.google.tagmanager',// 'angulartics.debug',
-                           'googlechart', 'angularFileUpload', 'angular-confirm', 'ngSanitize', 'swaggerUi',
+                           'googlechart', 'angularFileUpload', 'angular-confirm', 'ngSanitize', 'swaggerUi', 'ngCsv',
                            'app.resources', 'app.nav', 'app.compare', 'app.product', 'app.search', 'app.admin', 'app.overview', 'app.registration'])
         .config(['$routeProvider', function($routeProvider) {
             $routeProvider