+++ app/admin/components/userEditCtrl.js
@@ -25,6 +25,7 @@
                 vm.message = '';
                 if (!vm.acbId && !vm.atlId) {
                     vm.roles.push('ROLE_ADMIN');
+                    vm.roles.push('ROLE_CMS_STAFF');
                 }
                 if (!vm.atlId) {
                     vm.roles.push('ROLE_ACB_ADMIN');